+++ usher-common/usher-common-util/usher-common-util-scenario/src/main/java/team/opentech/usher/FitnessHandler.java
@@ -17,7 +17,7 @@ public interface FitnessHandler<T, E> {
      *
      * @return 适应度 范围为 0-1
      */
-    Float fitness(Individual<T, E> individual);
+    Double fitness(Individual<T, E> individual);
 
     /**
      * 计算适应度 适应度越趋近1则适应度越高
@@ -26,7 +26,7 @@ public interface FitnessHandler<T, E> {
      *
      * @return 适应度 范围为 0-1
      */
-    Float fitness(Individual<T, E> individual, T x);
+    Double fitness(Individual<T, E> individual, T x);
 
     /**
      * 计算适应度 适应度越趋近1则适应度越高
@@ -35,7 +35,7 @@ public interface FitnessHandler<T, E> {
      *
      * @return 适应度 范围为 0-1
      */
-    Float fitness(Individual<T, E> individual, T[] x);
+    Double fitness(Individual<T, E> individual, T[] x);
 
 
     /**
@@ -45,7 +45,7 @@ public interface FitnessHandler<T, E> {
      * @param percentage  要获取的比例 如果要获取前10% 则传入0.1
      * @param min         最少获取个数
      */
-    List<Individual<T, E>> findTopPercentage(List<Individual<T, E>> individuals, Float percentage, Integer min);
+    List<Individual<T, E>> findTopPercentage(List<Individual<T, E>> individuals, Double percentage, Integer min);
 
     /**
      * 计算平均适应度
@@ -55,7 +55,7 @@ public interface FitnessHandler<T, E> {
      *
      * @return
      */
-    Float fitnessByMean(List<Individual<T, E>> topPercentage, T[] params);
+    Double fitnessByMean(List<Individual<T, E>> topPercentage, T[] params);
 
     /**
      * 计算结果
