+++ usher-common/usher-common-util/usher-common-util-scenario/src/test/java/team/opentech/usher/core/heartDisease/TestHeartFunctionPopulation.java
@@ -10,25 +10,25 @@ import team.opentech.usher.core.AbstractPopulation;
  * @author uhyils <247452312@qq.com>
  * @date 文件创建日期 2024年03月27日 14时59分
  */
-public class TestHeartFunctionPopulation extends AbstractPopulation<Float[], Float> {
+public class TestHeartFunctionPopulation extends AbstractPopulation<Double[], Double> {
 
-    public TestHeartFunctionPopulation(Properties config, FitnessHandler<Float[], Float> fitnessHandler) {
+    public TestHeartFunctionPopulation(Properties config, FitnessHandler<Double[], Double> fitnessHandler) {
         super(config, fitnessHandler);
     }
 
-    public TestHeartFunctionPopulation(Integer k, Float initPercentage, Float crossPercentage, Float variationPercentage, FitnessHandler<Float[], Float> fitnessHandler) {
+    public TestHeartFunctionPopulation(Integer k, Double initPercentage, Double crossPercentage, Double variationPercentage, FitnessHandler<Double[], Double> fitnessHandler) {
         super(k, initPercentage, crossPercentage, variationPercentage, fitnessHandler);
     }
 
     @Override
-    protected Individual<Float[], Float> makeNewIndividual() {
+    protected Individual<Double[], Double> makeNewIndividual() {
         // todo 如何生成一个随机的新的个体
         //        return new TestHeartIndividual(firstDna(), secondDna());
         return null;
     }
 
     @Override
-    protected Float dealResults(List<Float> results) {
-        return (float) results.stream().mapToDouble(t -> t).average().getAsDouble();
+    protected Double dealResults(List<Double> results) {
+        return results.stream().mapToDouble(t -> t).average().getAsDouble();
     }
 }
