+++ usher-common/usher-common-util/usher-common-util-scenario/src/test/java/team/opentech/usher/core/twoDimensional/TestHistoryDataFunctionFitnessHandler.java
@@ -12,22 +12,22 @@ import team.opentech.usher.core.data.AbstractHistoryMeanData;
  * @author uhyils <247452312@qq.com>
  * @date 文件创建日期 2024年03月25日 16时24分
  */
-public class TestHistoryDataFunctionFitnessHandler extends AbstractDataFitnessHandler<Float, Float> {
+public class TestHistoryDataFunctionFitnessHandler extends AbstractDataFitnessHandler<Double, Double> {
 
     private Random random = new Random();
 
-    private Float[] dataKeys;
+    private Double[] dataKeys;
 
 
-    public TestHistoryDataFunctionFitnessHandler(Map<Float, Float> testData) {
+    public TestHistoryDataFunctionFitnessHandler(Map<Double, Double> testData) {
         super(testData);
-        this.dataKeys = testData.keySet().toArray(new Float[0]);
+        this.dataKeys = testData.keySet().toArray(new Double[0]);
         Arrays.sort(dataKeys);
     }
 
     @Override
-    public Float[] calculateResult(Float[] params) {
-        Float[] results = new Float[params.length];
+    public Double[] calculateResult(Double[] params) {
+        Double[] results = new Double[params.length];
         for (int i = 0; i < params.length; i++) {
             results[i] = calculateResult(params[i]);
         }
@@ -35,41 +35,41 @@ public class TestHistoryDataFunctionFitnessHandler extends AbstractDataFitnessHa
     }
 
     @Override
-    public Float calculateResult(Float param) {
+    public Double calculateResult(Double param) {
         return fittingIndividual.findResult(param);
     }
 
     @Override
-    protected AbstractHistoryData<Float, Float> makeFittingIndividual(Map<Float, Float> testData) {
-        return new AbstractHistoryMeanData<Float, Float>(testData, new Float[0]) {
+    protected AbstractHistoryData<Double, Double> makeFittingIndividual(Map<Double, Double> testData) {
+        return new AbstractHistoryMeanData<Double, Double>(testData, new Double[0]) {
 
             @Override
-            protected Float meanResult(Float param1, Float result1, Float param2, Float result2, Float param) {
-                float proportion = (param - param1) / (param2 - param1);
+            protected Double meanResult(Double param1, Double result1, Double param2, Double result2, Double param) {
+                double proportion = (param - param1) / (param2 - param1);
                 return proportion * (result2 - result1) + result1;
             }
 
             @Override
-            protected int compare(Float o1, Float o2) {
+            protected int compare(Double o1, Double o2) {
                 return o1.compareTo(o2);
             }
         };
     }
 
     @Override
-    protected Float doQuantifyGap(Float historicalResult, Float calculationResult) {
-        float abs = Math.abs(historicalResult - calculationResult);
+    protected Double doQuantifyGap(Double historicalResult, Double calculationResult) {
+        double abs = Math.abs(historicalResult - calculationResult);
         return historicalResult / (historicalResult + abs);
     }
 
     @NotNull
     @Override
-    protected Float[] makeTestParams() {
+    protected Double[] makeTestParams() {
         int size = 100;
-        Float[] randomX = new Float[size];
+        Double[] randomX = new Double[size];
         int randomSize = size / 2;
         for (int i = 0; i < randomSize; i++) {
-            float x = random.nextInt(20) + random.nextFloat();
+            double x = random.nextInt(20) + random.nextDouble();
             randomX[i] = x;
         }
         int findSize = size - randomSize;
